Task3

+++ Tyuiu.VolkovNS.Sprint5.Task3.V8.Lib/DataService.cs
@@ -12,7 +12,7 @@ namespace Tyuiu.VolkovNS.Sprint5.Task3.V8.Lib
             bool fileExists = fileinfo.Exists;
             if (fileExists)
                 File.Delete(path);
-            double res = Math.Round((x * x + 1) / (3 * x + 4),3);
+            double res = Math.Round(((double)x * (double)x + 1) / (3 * (double)x + 4),3);
             using (BinaryWriter writer = new BinaryWriter(File.Open(path, FileMode.OpenOrCreate), Encoding.UTF8))
             {
                 writer.Write(BitConverter.GetBytes(res));
